Redis persistence
Socket connections now persist to and retrieve from the redis database. There is further refactoring required, not least a mechanism for removing an entry when a user disconnects.

There is also a second redis client to support pub/sub so any changes to the redis keys is accompanied by a redis publish that the socket's version of the activity service is subscribed to and then accordingly handles and notifies users about.