Track what each socket is doing
The activity service now also tracks what each socket is doing - i.e., viewing or editing a case - and only allows one activity for each socket at any one time. When a socket disconnects, the activity is cleared out as it's considered that the user is no longer view/editing a case.

Also sorted out a bunch of lint errors and warnings.